+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.0.3 / WIP
+==================
+
+Bug fixes:
+* [OLMIS-4596](https://openlmis.atlassian.net/browse/OLMIS-4596): Fixed issue with sending a notification about nonfunctional CCE to users with home facility rights 
+
 1.0.2 / 2018-12-12
 ==================